latest build

+++ include/Add.h
@@ -36,17 +36,17 @@ namespace dnn
 			return GetDescriptionHeader();
 		}
 
-		size_t FanIn() const final override
+		UInt FanIn() const final override
 		{
 			return 1;
 		}
 
-		size_t FanOut() const final override 
+		UInt FanOut() const final override 
 		{
 			return 1;
 		}
 
-		void InitializeDescriptors(const size_t batchSize) final override
+		void InitializeDescriptors(const UInt batchSize) final override
 		{
 			if (InputLayer->DstMemDesc->data.ndims == 2)
 			{
@@ -97,7 +97,7 @@ namespace dnn
 #endif
 		}
 
-		void ForwardProp(const size_t batchSize, const bool training) final override
+		void ForwardProp(const UInt batchSize, const bool training) final override
 		{
 			if (training)
 			{
@@ -123,7 +123,7 @@ namespace dnn
 				{
 				case 2:
 				{
-					for_i(batchSize, threads, [=](size_t n)
+					for_i(batchSize, threads, [=](UInt n)
 					{
 						const auto start = n * size;
 						const auto end = start + part;
@@ -147,7 +147,7 @@ namespace dnn
 
 				case 3:
 				{
-					for_i(batchSize, threads, [=](size_t n)
+					for_i(batchSize, threads, [=](UInt n)
 					{
 						const auto start = n * size;
 						const auto end = start + part;
@@ -172,7 +172,7 @@ namespace dnn
 
 				case 4:
 				{
-					for_i(batchSize, threads, [=](size_t n)
+					for_i(batchSize, threads, [=](UInt n)
 					{
 						const auto start = n * size;
 						const auto end = start + part;
@@ -198,7 +198,7 @@ namespace dnn
 
 				default:
 				{
-					for_i(batchSize, threads, [=](size_t n)
+					for_i(batchSize, threads, [=](UInt n)
 					{
 						const auto start = n * size;
 						const auto end = start + part;
@@ -239,7 +239,7 @@ namespace dnn
 			}
 		}
 
-		void BackwardProp(const size_t batchSize) final override
+		void BackwardProp(const UInt batchSize) final override
 		{
 #ifdef DNN_LEAN
 			ZeroGradientMulti(batchSize);
@@ -297,7 +297,7 @@ namespace dnn
 				{
 				case 2:
 				{
-					for_i(batchSize, threads, [=](size_t n)
+					for_i(batchSize, threads, [=](UInt n)
 					{
 						const auto start = n * size;
 						const auto end = start + part;
@@ -326,7 +326,7 @@ namespace dnn
 
 				case 3:
 				{
-					for_i(batchSize, threads, [=](size_t n)
+					for_i(batchSize, threads, [=](UInt n)
 					{
 						const auto start = n * size;
 						const auto end = start + part;
@@ -360,7 +360,7 @@ namespace dnn
 
 				case 4:
 				{
-					for_i(batchSize, threads, [=](size_t n)
+					for_i(batchSize, threads, [=](UInt n)
 					{
 						const auto start = n * size;
 						const auto end = start + part;
@@ -400,7 +400,7 @@ namespace dnn
 
 				case 5:
 				{
-					for_i(batchSize, threads, [=](size_t n)
+					for_i(batchSize, threads, [=](UInt n)
 					{
 						const auto start = n * size;
 						const auto end = start + part;
@@ -445,7 +445,7 @@ namespace dnn
 
 				case 6:
 				{
-					for_i(batchSize, threads, [=](size_t n)
+					for_i(batchSize, threads, [=](UInt n)
 					{
 						const auto start = n * size;
 						const auto end = start + size;
@@ -465,7 +465,7 @@ namespace dnn
 
 				case 7:
 				{
-					for_i(batchSize, threads, [=](size_t n)
+					for_i(batchSize, threads, [=](UInt n)
 					{
 						const auto start = n * size;
 						const auto end = start + size;
@@ -486,7 +486,7 @@ namespace dnn
 
 				case 8:
 				{
-					for_i(batchSize, threads, [=](size_t n)
+					for_i(batchSize, threads, [=](UInt n)
 					{
 						const auto start = n * size;
 						const auto end = start + size;
@@ -507,7 +507,7 @@ namespace dnn
 				break;
 
 				default:
-					for_i(batchSize, threads, [=](size_t n)
+					for_i(batchSize, threads, [=](UInt n)
 					{
 						const auto start = n * size;
 						const auto end = start + size;
