+++ include/BatchNormActivationDropout.h
@@ -76,17 +76,17 @@ namespace dnn
 			return description;
 		}
 
-		size_t FanIn() const final override
+		UInt FanIn() const final override
 		{
 			return 1;
 		}
 
-		size_t FanOut() const final override
+		UInt FanOut() const final override
 		{
 			return 1;
 		}
 
-		void InitializeDescriptors(const size_t batchSize) final override
+		void InitializeDescriptors(const UInt batchSize) final override
 		{
 			if (InputLayer->DstMemDesc->data.ndims == 2)
 			{
@@ -111,7 +111,7 @@ namespace dnn
 			}
 		}
 
-		void SetBatchSize(const size_t batchSize) final override
+		void SetBatchSize(const UInt batchSize) final override
 		{
 			Layer::SetBatchSize(batchSize);
 
@@ -121,7 +121,7 @@ namespace dnn
 					NeuronsActive[n * PaddedCDHW + i] = Float(1);
 		}
 
-		void ForwardProp(const size_t batchSize, const bool training) final override
+		void ForwardProp(const UInt batchSize, const bool training) final override
 		{
 			const auto strideH = W * VectorSize;
 
@@ -131,7 +131,7 @@ namespace dnn
 				{
 					const auto partialHW = (HW / VectorSize) * VectorSize;
 
-					for_i(C, [=](size_t c)
+					for_i(C, [=](UInt c)
 					{
 						const auto runningMean = RunningMean[c];
 						const auto invStdDev = Float(1) / std::sqrt(RunningVariance[c] + Eps);
@@ -152,7 +152,7 @@ namespace dnn
 				}
 				else
 				{
-					for_i(PaddedC / VectorSize, [=](size_t c)
+					for_i(PaddedC / VectorSize, [=](UInt c)
 					{
 						const auto channelOffset = c * VectorSize;
 						const auto mapOffset = channelOffset * HW;
@@ -185,7 +185,7 @@ namespace dnn
 				{
 					const auto partialHW = (HW / VectorSize) * VectorSize;
 
-					for_i(C, [=](size_t c)
+					for_i(C, [=](UInt c)
 					{
 						auto vecMean = VecFloat(0);
 						auto mean = Float(0);
@@ -255,7 +255,7 @@ namespace dnn
 				}
 				else 
 				{
-					for_i(PaddedC / VectorSize, [=](size_t c)
+					for_i(PaddedC / VectorSize, [=](UInt c)
 					{
 						const auto channelOffset = c * VectorSize;
 						const auto mapOffset = channelOffset * HW;
@@ -321,7 +321,7 @@ namespace dnn
 			}
 		}
 
-		void BackwardProp(const size_t batchSize)  final override
+		void BackwardProp(const UInt batchSize)  final override
 		{
 #ifdef DNN_LEAN
 			ZeroGradient(batchSize);
@@ -333,7 +333,7 @@ namespace dnn
 			{
 				const auto partialHW = (HW / VectorSize) * VectorSize;
 				
-				for_i(C, [=](size_t c)
+				for_i(C, [=](UInt c)
 				{
 					const auto weightedInvStdDev = Scaling ? InvStdDev[c] * Weights[c] : InvStdDev[c];
 					const auto biases = Scaling && HasBias ? Biases[c] : Float(0);
@@ -410,7 +410,7 @@ namespace dnn
 			}
 			else
 			{
-				for_i(PaddedC / VectorSize, [=](size_t c)
+				for_i(PaddedC / VectorSize, [=](UInt c)
 				{
 					const auto channelOffset = c * VectorSize;
 					const auto mapOffset = channelOffset * HW;
@@ -424,7 +424,7 @@ namespace dnn
 					auto diffBeta = VecFloat(0);
 					auto diffSrc = VecFloat(0);
 
-					for (size_t n = 0; n < batchSize; ++n)
+					for (UInt n = 0; n < batchSize; ++n)
 					{
 						const auto offsetC = n * PaddedCDHW + mapOffset;
 						for (auto h = 0ull; h < H; ++h)
@@ -550,9 +550,9 @@ namespace dnn
 			return (2 * C * sizeof(Float)) + Layer::GetWeightsSize(persistOptimizer, optimizer);
 		}
 
-		size_t GetNeuronsSize(const size_t batchSize) const override
+		UInt GetNeuronsSize(const UInt batchSize) const override
 		{
-			size_t totalSize = 0;
+			UInt totalSize = 0;
 
 #ifndef DNN_LEAN
 			totalSize += PaddedCDHW * batchSize * sizeof(Float) * 3;
